refactor: reorganize header component folder structure
- Created `@/components/layout/header/menu` folder to separate menu-related files.
- Added `@/components/layout/header/menu/user` folder to organize user-specific menu components.
- Moved files from `@/components/layout/header` into the new structure for improved maintainability and clarity.
